rc2014: video work and prework
Play musical segments
Put the font into code3 (we need it post boot for VIDEO_MACCA)
Further work on the video drivers, udg stuff etc.

Still need to figure out how to sanely provide a UDG interface to the EF9345
40 column mode.

The VDP doesn't now need to pack the font away at the top of VDP space and
restore it but we keep the behaviour for consistency with MSX etc.