6809: Fix null handler signalling
Make it actually send a signal to the process (function 39), instead of
a chdir (function 10). For our syscall SWI handler the function should
be in B, the first argument in X and the remaining arguments on the stack.
For a normal system call there is the return address of the caller
on the stack between the SWI saved registers and the arguments.

Also tidy up the null vector scribble trap, which was unnecessarily
pushing an extra argument to the stack. ssig() takes a 16 bit and a
8 bit argument, thus via X and B.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>