mooh: Reboot via stub in internal memory
To imitate a hardware reset we need to disable the MMU
and also reset the MMU config for good measure. Then
jump into the ROM reset handler.

The last part must obviously run from internal RAM,
and either be copied there just-in-time, or slightly
simpler, be copied at start but therefore stored in the
common data address space where it is safe from overwriting.

As long as we have plenty of space at the bottom, we can
use the latter solution.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>