armm4: fix relocations
At least from testing and tracing this now seems right although the binaries
may be a bit bigger than needed and perhaps we can figure a way to dump the
junk.

There are still two failures - setdate and ls -l. Both appear to die in the
paths specific to using localtime() and at this point I suspect the bug is
not relocations (but who knows).

Now it works some of the changes here can possibly be unpicked but that can
wait until we know why setdate fails and are sure.

Also tweak armm4 to use the full malloc.