library: pieces for a "z80u" arch
This is a temporary ficticious z80 equivalent that allows us to start
test building the user space with the Fuzix compiler instead of SDCC. Once
things are stable it can move over to becoming the normal Z80 compiler and
we can then sync the compiler and assembler trees and tidy up the front end
to get the Z80 userspace proper being built with the Fuzix compilers.

Aside from the debug and the new pieces this will also need some improvements
to "ld" to generate relocatable Fuzix binaries. The ABI and other kernel
interfaces don't change (and indeed are already shared with the 8080 and 8085
builds using the new compiler)

Getting kernels building with the native compiler is a bigger project. For
unbanked it will need some small linker improvements but banked will require
a lot of work to replace the banking hacks currently in use with SDCC with
something resembling a proper banking model and needed linker changes.