ps: handle node size properly
Our kernel and user views may be a bit different which is ok providing the
relevant fields line up and the buffer is big enough. We check the buffer size
but did not handle the mis-sizing which could cause problems.

Calculate the offset using nodesize instead and also check for any remainders
(as suggested by veremenko-y on github).

Closes #1069 and #1070