genie-eg64: fix corner case in user copying
We can end up with a kernel object between 37E0 and 3FFF which is to
go to user space. However our user map has the video mapped in so
this is mapped over the kernel object we are copying.

Use a slightly customised clone of the user copier which maps user space
but not I/O when doing copying between spaces.