vdp: add support for targets that don't sufficiently waitstate the TMS I/O
Based on all the work from wigglethemouse and productiondave's testing
on the Z80 Retro discord

What seems to happen is that the Z80 at 10MHz without waitstates is enough
that
         out (c),l
         out (c),h

sometimes loses stuff and it needs a nop in the middle. We don't see this on
slower systems or on other fast ports as the other fast ports wait state the
TMS9918A so meet the timings.