setdate: support decimal encoded RTC devices
From: dennowiggle@github

The purpose of this issue is to document a change that may be helpful for
others.
I modified function rtcwrite() in the setdate application to support the real
time clock decimal format in addition to binary coded decimal.

I tested on an eZ80 system I am working on and could write both BCD and DEC
formats successfully. The eZ80 internal RTC supports both modes.

Here is the updated rtcwrite() function for setdate.c based on the 0.4 code
base.  (merged with 0.5 changes - AC)

Closes: #1122